### 2023-10-22 #5969 @editorialbot
[PRE REVIEW]: Surjectors: surjective normalizing flows for density estimation
**Submitting author:** @dirmeier (<a href="http://orcid.org/0000-0001-9795-3550">Simon Dirmeier</a>)
**Repository:** https://github.com/dirmeier/surjectors
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.0.3
