### 2023-11-24 #5969 @dirmeier
Hello, I think either generative modelling or neural density estimation which I would subsume under  _probabilistic deep learning_  or more generally _machine learning_.
