### 2024-01-07 #5969 @arfon
:wave: @VincentStimper @thomaspinder @animikhaich @sandeshkatakam @Uddiptaatwork – would any of you be willing to review this submission for JOSS? The submission under consideration is _Surjectors: surjective normalizing flows for density estimation_

Many thanks
Arfon
