### 2024-01-08 #6188 @editorialbot
[REVIEW]: Surjectors: surjective normalizing flows for density estimation
**Submitting author:** @dirmeier (<a href="http://orcid.org/0000-0001-9795-3550">Simon Dirmeier</a>)
**Repository:** https://github.com/dirmeier/surjectors
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.0.3
