### 2024-01-27 dirmeier/surjectors#25 @sandeshkatakam
Joss Review: Regarding the functionality of the Package
I have been trying to install and use the surjectors package, but I am facing a problem with importing sub-modules from surjectors. Specifically, I got this Module Not Found Error: surjectors.nn is not available

Can you guide me on how to work around this?\

Link to the Review Thread: https://github.com/openjournals/joss-reviews/issues/6188
