### 2024-01-27 dirmeier/surjectors#25 @dirmeier
Hey, thanks for testing. How did you try installing the package?
