### 2024-01-31 dirmeier/surjectors#26 @animikhaich
Joss Review: Example Usage
I installed Jax CPU, followed by Surjectors. A clean install inside a docker container.

Once installed, ran the example for constructing a simple normalizing flow as mentioned in the [Readme#examples](https://github.com/dirmeier/surjectors?tab=readme-ov-file#examples). I got the following error:

```py
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/usr/local/lib/python3.10/dist-packages/surjectors/bijectors/lu_linear.py", line 16, in __init__
    super().__init__(n_keep, None, None, "bijection", dtype)
  File "/usr/local/lib/python3.10/dist-packages/surjectors/surjectors/surjector.py", line 24, in __init__
    raise ValueError(
ValueError: 'kind' argument needs to be either of: inference_surjector/generative_surjector/bijector/surjector
```

It failed in the line: `transform = Chain([Slice(10, decoder_fn(10)), LULinear(5)])`

Link to the Review Thread: https://github.com/openjournals/joss-reviews/issues/6188
