### 2024-02-01 aba2531
Update documentation, add test
