+++ docs/index.rst
@@ -26,8 +26,8 @@ Example
 You can, for instance, construct a simple normalizing flow like this:
 
     >>> import distrax
-    >>> from jax import numpy as jnp
-    >>> from surjectors import Slice, LULinear, Chain
+    >>> import haiku as hk
+    >>> from jax import numpy as jnp, random as jr
     >>> from surjectors import TransformedDistribution
     >>> from surjectors.nn import make_mlp
     >>>
@@ -38,9 +38,19 @@ You can, for instance, construct a simple normalizing flow like this:
     >>>         return distrax.Independent(distrax.Normal(means, jnp.exp(log_scales)))
     >>>     return _fn
     >>>
-    >>> base_distribution = distrax.Normal(jnp.zeros(5), jnp.ones(1))
-    >>> transform = Chain([Slice(10, decoder_fn(10)), LULinear(5)])
-    >>> pushforward = TransformedDistribution(base_distribution, transform)
+    >>> @hk.without_apply_rng
+    >>> @hk.transform
+    >>> def flow(x):
+    >>>     base_distribution = distrax.Independent(
+    >>>         distrax.Normal(jnp.zeros(5), jnp.ones(5)), 1
+    >>>     )
+    >>>     transform = Chain([Slice(5, decoder_fn(5)), LULinear(5)])
+    >>>     pushforward = TransformedDistribution(base_distribution, transform)
+    >>>     return pushforward.log_prob(x)
+    >>>
+    >>> x = jr.normal(jr.PRNGKey(1), (1, 10))
+    >>> params = flow.init(jr.PRNGKey(2), x)
+    >>> lp = flow.apply(params, x)
 
 The flow is constructed using three objects: a base distribution, a transformation, and a transformed distribution.
 
