### 2024-02-01 6ca6602
Update documentation, add test (#27)
* Update documentation, add test
* Remove backticks in lieu of quotes
