### 2024-02-01 dirmeier/surjectors#26 @dirmeier
Hi, my bad, I hadn't updated the pip version to `v0.3.0`. Assuming you are running version `v0.3.0`, the examples should run (at least the GitHub actions run and the documentation builds).

I updated the example in the README to make it more friendly for users w/o Haiku background. Could you test this:

```bash
pip install surjectors==0.3.0
```

and then

```python
import distrax
import haiku as hk
from jax import numpy as jnp, random as jr
from surjectors import Slice, LULinear, Chain
from surjectors import TransformedDistribution
from surjectors.nn import make_mlp

def decoder_fn(n_dim):
    def _fn(z):
        params = make_mlp([32, 32, n_dim * 2])(z)
        means, log_scales = jnp.split(params, 2, -1)
        return distrax.Independent(distrax.Normal(means, jnp.exp(log_scales)))
    return _fn

@hk.without_apply_rng
@hk.transform
def flow(x):
    base_distribution = distrax.Independent(
        distrax.Normal(jnp.zeros(5), jnp.ones(5)), 1
    )
    transform = Chain([Slice(5, decoder_fn(5)), LULinear(5)])
    pushforward = TransformedDistribution(base_distribution, transform)
    return pushforward.log_prob(x)

x = jr.normal(jr.PRNGKey(1), (1, 10))
params = flow.init(jr.PRNGKey(2), x)
lp = flow.apply(params, x)
print(lp)
```

Thanks and cheers,
Simon
