### 2024-02-01 dirmeier/surjectors#26 @dirmeier
I think the example in the README is not very illustrative though. The ones in [examples](https://github.com/dirmeier/surjectors/tree/main/examples) are better.
