+++ docs/requirements.txt
@@ -4,6 +4,7 @@ ipython
 matplotlib
 nbsphinx
 pandas
+scikit-learn
 seaborn
 session_info
 sphinx
