### 2024-02-01 f8eaa7d
Change docu style (#28)
* Update documentation styles

* Add them colors

* Rerun notebooks
