+++ surjectors/__init__.py
@@ -2,7 +2,7 @@
 surjectors: Surjection layers for density estimation with normalizing flows
 """
 
-__version__ = "0.3.0"
+__version__ = "0.3.1"
 
 from surjectors._src.bijectors.lu_linear import LULinear
 from surjectors._src.bijectors.masked_autoregressive import MaskedAutoregressive
