+++ docs/requirements.txt
@@ -1,15 +1,15 @@
 -e .
-furo
 ipykernel
 ipython
 matplotlib
 nbsphinx
 pandas
+scikit-learn
 seaborn
 session_info
 sphinx
 sphinx-autobuild
-sphinx-book-theme
+sphinx-book-theme>=1.1.0
 sphinx-copybutton
 sphinx-math-dollar
 sphinx_autodoc_typehints
