+++ pyproject.toml
@@ -15,10 +15,11 @@ classifiers = [
     "Intended Audience :: Science/Research",
     "License :: OSI Approved :: Apache Software License",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
 ]
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 dependencies = [
     "distrax>=0.1.4",
     "dm-haiku>=0.0.10",
@@ -60,7 +61,7 @@ test = 'pytest -v --cov=./surjectors --cov-report=xml surjectors'
 [tool.black]
 line-length = 80
 extend-ignore = "E203"
-target-version = ['py39']
+target-version = ['py310']
 exclude = '''
 /(
     \.eggs
