### 2024-02-02 dirmeier/surjectors#26 @dirmeier
Thanks! I'll made a PR #29 with the fix. Ill merge it after the review is done (I dont wanna add too many non-functional commits).
