### 2024-02-05 #6188 @animikhaich
Hi @arfon, The Review is completed from my end. We are good to go! Please let me know if anything else is needed.

Please refer to the author's comments for the aforementioned documentation fix: https://github.com/dirmeier/surjectors/issues/26#issuecomment-1923375045
