### 2024-02-19 c006920
Joss paper (#24)
