+++ paper/paper.bib
@@ -22,13 +22,12 @@
   year={2023}
 }
 
-@article{dirmeier2023ood,
+@article{dirmeier2023uncertainty,
   title={Uncertainty quantification and out-of-distribution detection using surjective normalizing flows},
   author={Dirmeier, Simon and Hong, Ye and Xin, Yanan and Perez-Cruz, Fernando},
-  journal={In preparation},
+  journal={arXiv preprint arXiv:2311.00377},
   year={2023}
 }
-
 
 @article{hoffman2019neutra,
   title={Neutra-lizing bad geometry in hamiltonian monte carlo using neural transport},
@@ -141,3 +140,4 @@
   booktitle = 	 {Proceedings of the 35th International Conference on Machine Learning},
   year = 	 {2018}
 }
+
