### 2024-02-22 #6188 @dirmeier
> @dirmeier – could you update the license of the Zenodo record to be the same as the software please?

@arfon , done! Thanks!
